+++ src/deprecated.jl
@@ -130,3 +130,11 @@ function MultiSiteOperator(args...)
     )
     return multisite_operator(args...)
 end
+
+export ProgressBar
+function ProgressBar(args...; kwargs...)
+    # Use error instead of depwarn, since ProgressBar and Progress have different arguments
+    return error(
+        "`ProgressBar` is deprecated and will be removed in next major release. Use `Progress` from `ProgressMeter.jl` instead.",
+    )
+end